requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
diff-match-patch==20241021
Django==5.0.7
django-cors-headers==4.6.0
django-filter==24.3
django-import-export==4.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
Markdown==3.6
mysql-connector-python==9.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tablib==3.7.0
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
